Gesture-Generation-from-Multimodal-Context/requirements.txt
torch>=1.1.0
numpy
scipy
scikit-learn
matplotlib
librosa
tensorboard>=1.14
future
lmdb==0.96
tqdm
fasttext
configargparse
soundfile
pygame
google-cloud-texttospeech==1.0.1
